+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Unreleased
 * Added new async method: ``Instrument.publisher()``
 * Added new async method: ``Instrument.share_holders()``
 * Added new async method: ``Instrument.share_holder_history()``
+* Added new async method: ``Instrument.messages()``
 * Added new async function: ``instruments.share_holder_companies()``
 * Improved type annotations
 * Deprecated the following methods which rely on the old tsetmc site and have new alternatives:
@@ -22,6 +23,7 @@ Unreleased
   * ``Insturment.introduction``. Use ``publisher`` instead.
   * ``Insturment.holders``. Use ``share_holders`` instead.
   * ``Insturment.holder``. Use ``share_holder_history`` or ``share_holder_companies`` instead.
+  * ``Insturment.ombud_messages``. Use ``messages`` instead.
 
 * ``InstrumentOnDate.client_types()`` is deprecated, use ``client_type()`` instead
 * Updated dataset